% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_isolates.R
\name{connect_isolates_ig}
\alias{connect_isolates_ig}
\title{Connect isolated nodes to themselves in an igraph graph}
\usage{
connect_isolates_ig(ntwk_ig)
}
\arguments{
\item{ntwk_ig}{the iGraph network object}
}
\value{
the iGraph network object with any isolates self-connected
}
\description{
Connect isolated nodes to themselves in an igraph graph
}
\examples{
\dontrun{
flo_ig <- connect_isolates_sna(flo_ig)
}
}
