% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_isolates.R
\name{connect_isolates_sna}
\alias{connect_isolates_sna}
\title{Connect isolated nodes to themselves in a statnet graph}
\usage{
connect_isolates_sna(ntwk_sna)
}
\arguments{
\item{ntwk_sna}{the statnet network object}
}
\value{
the statnet network object with any isolates self-connected
}
\description{
Connect isolated nodes to themselves in a statnet graph
}
\examples{
\dontrun{
flo_stat <- connect_isolates_sna(flo_stat)
}
}
