% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntwk_info.R
\name{ntwk_info_sna}
\alias{ntwk_info_sna}
\title{Return information about a statnet network object}
\usage{
ntwk_info_sna(ntwk_sna)
}
\arguments{
\item{ntwk_sna}{the statnet network object}
}
\value{
a list of network information
}
\description{
Return information about a statnet network object
}
\examples{
x <- netwk_info_sna(flomarr.sna)
flo_dir <- x$is_directed
}
