% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ntwk_pairs.R
\name{ntwk_vector_pairs}
\alias{ntwk_vector_pairs}
\title{Title}
\usage{
ntwk_vector_pairs(ig_nodes, sna_nodes, col_name)
}
\arguments{
\item{ig_nodes}{a list of igraph nodes from \code{ntwk_info_ig}}

\item{sna_nodes}{a list of statnet nodes from \code{ntwk_info_sna}}

\item{col_name}{the name of the column to create data pairs for}
}
\value{
a list of paired vectors of values, suitable to pass to \code{plot_ntwk_vector_pairs}
}
\description{
Title
}
