% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ntwk_pairs.R
\name{plot_ntwk_vector_pairs}
\alias{plot_ntwk_vector_pairs}
\title{Title}
\usage{
plot_ntwk_vector_pairs(ntwk_list)
}
\arguments{
\item{ntwk_list}{a list of vector pairs from \code{newk_vector_pairs}}
}
\value{

}
\description{
Title
}
